NU-1858 edit properties option
