+++ designer/client/src/actions/actionTypes.ts
@@ -41,4 +41,5 @@ export type ActionTypes =
     | "PROCESS_VERSIONS_LOADED"
     | "UPDATE_BACKEND_NOTIFICATIONS"
     | "MARK_BACKEND_NOTIFICATION_READ"
-    | "ARCHIVED";
+    | "ARCHIVED"
+    | "EDIT_PROPERTIES";
