+++ designer/client/src/actions/nk/calculateProcessAfterChange.ts
@@ -1,20 +1,8 @@
-import NodeUtils from "../../components/graph/NodeUtils";
-import { fetchProcessDefinition } from "./processDefinitionData";
 import { getProcessDefinitionData } from "../../reducers/selectors/settings";
 import { mapProcessWithNewNode, replaceNodeOutputEdges } from "../../components/graph/utils/graphUtils";
-import { alignFragmentWithSchema } from "../../components/graph/utils/fragmentSchemaAligner";
-import { Edge, NodeType, ScenarioGraph, ProcessDefinitionData, ScenarioGraphWithName } from "../../types";
+import { Edge, NodeType, ScenarioGraphWithName } from "../../types";
 import { ThunkAction } from "../reduxTypes";
 import { Scenario } from "../../components/Process/types";
-
-function alignFragmentsNodeWithSchema(scenarioGraph: ScenarioGraph, processDefinitionData: ProcessDefinitionData): ScenarioGraph {
-    return {
-        ...scenarioGraph,
-        nodes: scenarioGraph.nodes.map((node) => {
-            return node.type === "FragmentInput" ? alignFragmentWithSchema(processDefinitionData, node) : node;
-        }),
-    };
-}
 
 export function calculateProcessAfterChange(
     scenario: Scenario,
@@ -22,23 +10,7 @@ export function calculateProcessAfterChange(
     after: NodeType,
     outputEdges: Edge[],
 ): ThunkAction<Promise<ScenarioGraphWithName>> {
-    return async (dispatch, getState) => {
-        if (NodeUtils.nodeIsProperties(after)) {
-            const processDefinitionData = await dispatch(fetchProcessDefinition(scenario.processingType, scenario.isFragment));
-            const processWithNewFragmentSchema = alignFragmentsNodeWithSchema(scenario.scenarioGraph, processDefinitionData);
-            // TODO: We shouldn't keep scenario name in properties.id - it is a top-level scenario property
-            if (after.id !== before.id) {
-                dispatch({ type: "PROCESS_RENAME", name: after.id });
-            }
-
-            const { id, ...properties } = after;
-
-            return {
-                processName: after.id,
-                scenarioGraph: { ...processWithNewFragmentSchema, properties },
-            };
-        }
-
+    return async (_, getState) => {
         let changedProcess = scenario.scenarioGraph;
         if (outputEdges) {
             const processDefinitionData = getProcessDefinitionData(getState());
@@ -54,7 +26,7 @@ export function calculateProcessAfterChange(
         }
 
         return {
-            processName: scenario.scenarioGraph.properties.id || scenario.name,
+            processName: scenario.name,
             scenarioGraph: mapProcessWithNewNode(changedProcess, before, after),
         };
     };
